+++ src/libse/Common/Settings/CompareSettings.cs
@@ -1,0 +1,16 @@
+namespace Nikse.SubtitleEdit.Core.Common
+{
+    public class CompareSettings
+    {
+        public bool ShowOnlyDifferences { get; set; }
+        public bool OnlyLookForDifferenceInText { get; set; }
+        public bool IgnoreLineBreaks { get; set; }
+        public bool IgnoreWhitespace { get; set; }
+        public bool IgnoreFormatting { get; set; }
+
+        public CompareSettings()
+        {
+            OnlyLookForDifferenceInText = true;
+        }
+    }
+}
