+++ src/libse/Common/Settings/FcpExportSettings.cs
@@ -1,0 +1,22 @@
+using System.Drawing;
+
+namespace Nikse.SubtitleEdit.Core.Common
+{
+    public class FcpExportSettings
+    {
+        public string FontName { get; set; }
+        public int FontSize { get; set; }
+        public string Alignment { get; set; }
+        public int Baseline { get; set; }
+        public Color Color { get; set; }
+
+        public FcpExportSettings()
+        {
+            FontName = "Lucida Sans";
+            FontSize = 36;
+            Alignment = "center";
+            Baseline = 29;
+            Color = Color.WhiteSmoke;
+        }
+    }
+}
