+++ src/libse/Common/Settings/MultipleSearchAndReplaceGroup.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Nikse.SubtitleEdit.Core.Common
+{
+    public class MultipleSearchAndReplaceGroup
+    {
+        public string Name { get; set; }
+        public bool Enabled { get; set; }
+        public List<MultipleSearchAndReplaceSetting> Rules { get; set; }
+    }
+}
