+++ src/libse/Common/Settings/MultipleSearchAndReplaceSetting.cs
@@ -1,0 +1,11 @@
+namespace Nikse.SubtitleEdit.Core.Common
+{
+    public class MultipleSearchAndReplaceSetting
+    {
+        public bool Enabled { get; set; }
+        public string FindWhat { get; set; }
+        public string ReplaceWith { get; set; }
+        public string SearchType { get; set; }
+        public string Description { get; set; }
+    }
+}
