+++ src/libse/Common/Settings/WordListSettings.cs
@@ -1,0 +1,15 @@
+namespace Nikse.SubtitleEdit.Core.Common
+{
+    public class WordListSettings
+    {
+        public string LastLanguage { get; set; }
+        public string NamesUrl { get; set; }
+        public bool UseOnlineNames { get; set; }
+
+        public WordListSettings()
+        {
+            LastLanguage = "en-US";
+            NamesUrl = "https://raw.githubusercontent.com/SubtitleEdit/subtitleedit/main/Dictionaries/names.xml";
+        }
+    }
+}
